Move server data to a different file.
