+++ app/FakeServerData.js
@@ -1,0 +1,8 @@
+const SERVER_DATA = {
+  message: 'El torito les desea felices pascuas',
+  facilityName: 'El Torito',
+  reservationCovers: 2,
+  reservationDate: new Date()
+};
+
+export default SERVER_DATA;
